# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MealFlow API
  description: India-focused family meal planning app API
  version: 0.9.0
  license:
    name: MIT
servers:
  - url: http://localhost:3458
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mf_sid
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Recipe:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        servings:
          type: integer
        prep_time:
          type: integer
        cook_time:
          type: integer
        cuisine:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        is_favorite:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredient'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        nutrition:
          $ref: '#/components/schemas/NutritionSummary'
    RecipeIngredient:
      type: object
      properties:
        id:
          type: integer
        ingredient_id:
          type: integer
        quantity:
          type: number
        unit:
          type: string
        notes:
          type: string
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        recipe_count:
          type: integer
...
    NutritionSummary:
      type: object
      properties:
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
security:
  - cookieAuth: []
paths:
...
  /api/recipes:
    get:
      tags: [recipes]
      summary: List recipes (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: cuisine
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: favorite
          in: query
          schema:
            type: string
            enum: ['1']
        - name: q
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created_at, updated_at]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated recipe list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Recipe'
    post:
      tags: [recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                servings:
                  type: integer
                prep_time:
                  type: integer
                cook_time:
                  type: integer
                cuisine:
                  type: string
                difficulty:
                  type: string
                ingredients:
                  type: array
                  items:
                    type: object
                    properties:
                      ingredient_id:
                        type: integer
                      quantity:
                        type: number
                      unit:
                        type: string
                tags:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
  /api/recipes/{id}:
    get:
      tags: [recipes]
      summary: Get a recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                servings:
                  type: integer
      responses:
        '200':
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
    delete:
      tags: [recipes]
      summary: Delete a recipe (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /api/recipes/search:
    get:
      tags: [recipes]
      summary: Full-text search recipes
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: cuisine
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
...